requirements.txt
# Core data processing
pandas==2.2.2
numpy==1.26.4
pyarrow==21.0.0
python-dateutil==2.9.0.post0

# API and network
requests==2.32.4
aiohttp==3.9.5
tenacity==9.1.2

# Excel processing
openpyxl==3.1.3
xlsxwriter==3.2.5

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.30.0

# Google integration
gspread==6.2.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-api-python-client>=2.100.0
google-auth>=2.28.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.2.0

# Utilities
python-dotenv==1.1.1
colorama>=0.4.6
pydantic>=2.5.0 
Jinja2
PyJWT>=2.8.0